+++ common/logging/src/main/kotlin/net/corda/common/logging/Constants.kt
@@ -8,4 +8,4 @@ package net.corda.common.logging
  * the generated file does not need to be committed to source control (originally added to source control for ease of use)
  */
 
-internal const val CURRENT_MAJOR_RELEASE = "4.3-SNAPSHOT"
+internal const val CURRENT_MAJOR_RELEASE = "4.3-RC01"