+++ node/src/main/kotlin/net/corda/node/migration/VaultStateMigration.kt
@@ -16,6 +16,7 @@ import net.corda.core.serialization.internal.effectiveSerializationEnv
 import net.corda.core.utilities.contextLogger
 import net.corda.node.internal.DBNetworkParametersStorage
 import net.corda.node.services.identity.PersistentIdentityService
+import net.corda.node.services.keys.BasicHSMKeyManagementService
 import net.corda.node.services.persistence.DBTransactionStorage
 import net.corda.node.services.persistence.NodeAttachmentService
 import net.corda.node.services.vault.NodeVaultService
@@ -139,12 +140,14 @@ object VaultMigrationSchemaBuilder {
                 mappedTypes = listOf(
                         DBTransactionStorage.DBTransaction::class.java,
                         PersistentIdentityService.PersistentPublicKeyHashToCertificate::class.java,
-                PersistentIdentityService.PersistentPartyToPublicKeyHash::class.java,
+                        PersistentIdentityService.PersistentPartyToPublicKeyHash::class.java,
                         PersistentIdentityService.PersistentPublicKeyHashToParty::class.java,
+                        PersistentIdentityService.PersistentHashToPublicKey::class.java,
+                        BasicHSMKeyManagementService.PersistentKey::class.java,
                         NodeAttachmentService.DBAttachment::class.java,
                         DBNetworkParametersStorage.PersistentNetworkParameters::class.java
-                ))
-}
+        )
+)
 
 /**
  * Provides a mechanism for iterating through all persistent vault states.